# Idempotent indexing in range[IdOffsetRange]