After this PR:

```julia
julia> r = OffsetArrays.IdOffsetRange(3:5, -1)
OffsetArrays.IdOffsetRange(2:4)

julia> (1:4)[r]
2:4 with indices 0:2

julia> axes((1:4)[r]) == axes(r)
true
```

On master:

```julia
julia> (1:4)[r]
2:4
```

As a consequence, 

on master:

```julia
julia> a = zeros(2:4);

julia> r = 1:4;

julia> a[axes(a,1)] .= r[axes(a,1)]
ERROR: DimensionMismatch("array could not be broadcast to match destination")
```

after this PR:

```julia
julia> a = zeros(2:4);

julia> r = 1:4;

julia> a[axes(a,1)] .= r[axes(a,1)]
3-element view(OffsetArray(::Array{Float64,1}, 2:4), 2:4) with eltype Float64 with indices 2:4:
 2.0
 3.0
 4.0
```