@timholy 2020-10-09
It's a bit suspicious that the bodies of all the new methods are identical. Can you instead modify the generic `AbstractRange` methods above to achieve the same thing?