@jishnub 2020-10-10
It appears to be a bit difficult to achieve this in general without introducing ambiguities, as some of the range types have specific indexing behaviors defined in `Base`. Eg: 

```julia
function getindex(r::StepRange, s::AbstractRange{<:Integer})
```

and 

```julia
function getindex(r::StepRangeLen{T,<:TwicePrecision,<:TwicePrecision}, s::OrdinalRange{<:Integer}) where T
```

presumably indicate that we should specialize on the second argument?