@timholy 2020-10-10
Ah, yeah, good point. At some point we probably want to have some tests for this kind of thing that introduce a novel range object (could use https://github.com/JuliaArrays/CustomUnitRanges.jl to make that easy), just to make sure we go as far as we can with generic operations.