@timholy 2020-10-10
I have no other objections, though. This clearly improves the handling of range values and range indices :+1:. So merge at will.