@jishnub 2020-10-10
Yes indeed. The present approach is patchwork at best, and a new range type introduced will not support idempotent indexing with an `IdOffsetRange`. I wonder what's the best approach to address this?

For example:

```julia
julia> struct MyUnitRange <: AbstractUnitRange{Int}
       start :: Int
       stop :: Int
       end

julia> Base.first(r::MyUnitRange) = r.start

julia> Base.last(r::MyUnitRange) = r.stop

julia> s = MyUnitRange(2,3);

julia> r = OffsetArrays.IdOffsetRange(4:5, -3)
OffsetArrays.IdOffsetRange(1:2)

julia> axes(r)
(OffsetArrays.IdOffsetRange(-2:-1),)

julia> s[r]
2:3

julia> axes(s[r])
(Base.OneTo(2),)
```

Ideally we should be able to define 

```julia
getindex(r::AbstractRange, inds::IdOffsetRange)
```
however the problem with this is that a type that defines 

```julia
getindex(r::CustomRange, inds::AbstractUnitRange{<:Integer})
```
will run into ambiguities.