@johnnychen94 2020-10-10
I'm still a little confused about how things are working under the hood, but tests look good. Very concise a patch, I agree that we need tests to make sure things don't get break unexpectedly.