@timholy 2020-10-10
I suspect it's a sign we may need better "inner methods" and implement the core range algorithms with them...although ranges are already pretty close to the metal. I suspect this is a multistage process, where we get this working and then rethink range implementation in Base.