@jishnub 2020-10-10
I've added methods that return an `AbstractUnitRange` instead of an `OffsetArray` when a `UnitRange{<:Integer}` is indexed with an `IdOffstRange` or an `IdentityUnitRange`. This is because the result may be further used in indexing, and indexing with ranges is faster than that with vectors.

before
```julia
julia> a = collect(1:10);

julia> ur = 3:6;

julia> idr = Base.IdentityUnitRange(2:4);

julia> @btime $a[$ur[$idr]];
  98.796 ns (1 allocation: 112 bytes)

julia> idor = OffsetArrays.IdOffsetRange(1:3, 1);

julia> @btime $a[$ur[$idor]];
  102.616 ns (1 allocation: 112 bytes)
```

after
```julia
julia> @btime $a[$ur[$idr]];
  76.405 ns (1 allocation: 112 bytes)

julia> @btime $a[$ur[$idor]];
  84.349 ns (1 allocation: 112 bytes)
```